<feature> advanced theory
This commit adds the blues scales (major and minor) as well as some advanced chords in a dropdown menu when a chord is selected.
The default chord display was updated to show notes in the current chord on top of the notes of the current scale. This allows the user to see how a chord fits into the scale of the selected key.